- Fix rounding precision - Methods get/set/add/sub - Methods isAfter/isBefore/isBetween
