+++ tests/CarbonImmutable/SettersTest.php
@@ -351,6 +351,8 @@ class SettersTest extends AbstractTestCase
     public function testMidDayAtSetter()
     {
         $d = Carbon::now();
+        $d->setMidDayAt(11);
+        $this->assertSame(11, $d->getMidDayAt());
         $d->setMidDayAt(12);
         $this->assertSame(12, $d->getMidDayAt());
     }
